pretyped int to float

+++ Assets/FireballDMG.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class FireballDMG : MonoBehaviour
 {
-    public int damage = 2;
+    public float damage = 2;
 
     void OnTriggerEnter(Collider other)
     {
