+++ Assets/ManaSystem.cs
@@ -6,11 +6,11 @@ using UnityEngine.UI;
 public class ManaSystem : MonoBehaviour
 {
     public Slider manaSlider;
-    public int maxMana = 100;
-    public int currentMana;
-    public int fireballManaCost = 10;
-    public int flamethrowerCost = 2;
-    public int manaRegen = 1;
+    public float maxMana = 100;
+    public float currentMana;
+    public float fireballManaCost = 10;
+    public float flamethrowerCost = 2;
+    public float manaRegen = 1;
     public float regenInterval = 0.2f;
     public float vibrationDuration = 0.5f;
     public float vibrationMagnitude = 7.1f;
@@ -83,7 +83,7 @@ public class ManaSystem : MonoBehaviour
         StopSpendingManaPerSecond(); // Ensure it stops when mana is insufficient
     }
 
-    public bool SpendMana(int amount)
+    public bool SpendMana(float amount)
     {
         if (currentMana >= amount)
         {
